Create diskget.c
Implemented a program to copy a file from the current Linux directory into the file system, either at the root directory or a given sub-directory. Outputs a 'File not found' message if the specified file is not found in the current directory. Invoked with the syntax './diskput test.img foo.txt /sub_dir/foo3.txt'. Only works for grabbing information if there is enough continuos memory